pyproject.toml
[tool.poetry]
name = "lab3"
version = "0.2.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
requests = "^2.30.0"
joblib = "^1.2.0"
#scipy = "^1.10.1"
scikit-learn = "1.2.2"
fastapi-cache2 = "0.1.9"
redis = "^4.6.0"
#memcache = "^0.12.0"
#aiobotocore = "^2.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
black = "^23.3"
isort = "^5.12"
httpx = "^0.24"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
